initial upload
uploading the files for the dark mode settings to make apple-music-for-linux use a black background and white text. I am still working on figuring out how to change that awful white to something like a blue. If you know how please let me know. Until then enjoy this big step in the right direction. Just copy and paste the folder over your current install and restart the app.
If you want to look closer, it's just a few folders, changing the gtk3 settings, and adding the missing setting for gtk4.